+++ lib/assent/strategies/line.ex
@@ -2,6 +2,13 @@ defmodule Assent.Strategy.LINE do
   @moduledoc """
   LINE Login OpenID Connect Strategy.
 
+  ## Usage
+
+    config = [
+      client_id: "REPLACE_WITH_CLIENT_ID",
+      client_secret: "REPLACE_WITH_CLIENT_SECRET"
+    ]
+
   See `Assent.Strategy.OIDC` for more.
   """
 
@@ -15,7 +22,4 @@ defmodule Assent.Strategy.LINE do
       id_token_signed_response_alg: "HS256"
     ]
   end
-
-  @impl true
-  def normalize(_config, user), do: {:ok, user}
 end