+++ test/assent/http_adapter/mint_test.exs
@@ -2,51 +2,71 @@ defmodule Assent.HTTPAdapter.MintTest do
   use ExUnit.Case
   doctest Assent.HTTPAdapter.Mint
 
-  alias ExUnit.CaptureLog
   alias Mint.TransportError
   alias Assent.HTTPAdapter.{HTTPResponse, Mint}
 
-  @wrong_host_certificate_url "https://wrong.host.badssl.com"
-  @hsts_certificate_url "https://hsts.badssl.com"
-  @unreachable_http_url "http://localhost:8888/"
-
   describe "request/4" do
     test "handles SSL" do
-      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, @hsts_certificate_url, nil, [])
-      assert {:error, %TransportError{reason: {:tls_alert, {:handshake_failure, _error}}}} = Mint.request(:get, @wrong_host_certificate_url, nil, [])
+      TestServer.start(scheme: :https)
+      TestServer.add("/", via: :get)
 
-      # For OTP 24 "Authenticity is not established by certificate path validation" warning
-      CaptureLog.capture_log(fn ->
-        assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, @wrong_host_certificate_url, nil, [], transport_opts: [verify: :verify_none])
-      end)
+      mint_opts = [transport_opts: [cacerts: TestServer.x509_suite().cacerts], protocols: [:http1]]
 
-      assert {:error, %TransportError{reason: :econnrefused}} = Mint.request(:get, @unreachable_http_url, nil, [])
+      assert {:ok, %HTTPResponse{status: 200, body: "HTTP/1.1"}} = Mint.request(:get, TestServer.url(), nil, [], mint_opts)
+    end
+
+    test "handles SSL with bad certificate" do
+      TestServer.start(scheme: :https)
+
+      bad_host_url = TestServer.url(host: "bad-host.localhost")
+      mint_opts = [transport_opts: [cacerts: TestServer.x509_suite().cacerts]]
+
+      assert {:error, %TransportError{reason: {:tls_alert, {:handshake_failure, _error}}}} = Mint.request(:get, bad_host_url, nil, [], mint_opts)
+    end
+
+    test "handles SSL with bad certificate and no verification" do
+      TestServer.start(scheme: :https)
+      TestServer.add("/", via: :get)
+
+      bad_host_url = TestServer.url(host: "bad-host.localhost")
+      mint_opts = [transport_opts: [cacerts: TestServer.x509_suite().cacerts, verify: :verify_none]]
+
+      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, bad_host_url, nil, [], mint_opts)
     end
 
     if :crypto.supports()[:curves] do
       test "handles http/2" do
-        assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, "https://http2.golang.org/", nil, [])
+        TestServer.start(scheme: :https)
+        TestServer.add("/", via: :get)
+
+        mint_opts = [transport_opts: [cacerts: TestServer.x509_suite().cacerts]]
+
+        assert {:ok, %HTTPResponse{status: 200, body: "HTTP/2"}} = Mint.request(:get, TestServer.url(), nil, [], mint_opts)
       end
     else
       IO.warn("No support curve algorithms, can't test in #{__MODULE__}")
     end
 
-    test "handles query in URL" do
-      bypass = Bypass.open()
+    test "handles unreachable host" do
+      TestServer.start()
+      url = TestServer.url()
+      TestServer.stop()
 
-      Bypass.expect_once(bypass, "GET", "/get", fn conn ->
+      assert {:error, %TransportError{reason: :econnrefused}} = Mint.request(:get, url, nil, [])
+    end
+
+    test "handles query in URL" do
+      TestServer.add("/get", via: :get, to: fn conn ->
         assert conn.query_string == "a=1"
 
         Plug.Conn.send_resp(conn, 200, "")
       end)
 
-      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, "http://localhost:#{bypass.port}/get?a=1", nil, [])
+      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:get, TestServer.url("/get?a=1"), nil, [])
     end
 
     test "handles POST" do
-      bypass = Bypass.open()
-
-      Bypass.expect_once(bypass, "POST", "/post", fn conn ->
+      TestServer.add("/post", via: :post, to: fn conn ->
         {:ok, body, conn} = Plug.Conn.read_body(conn, [])
         params = URI.decode_query(body)
 
@@ -57,7 +77,7 @@ defmodule Assent.HTTPAdapter.MintTest do
         Plug.Conn.send_resp(conn, 200, "")
       end)
 
-      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:post, "http://localhost:#{bypass.port}/post", "a=1&b=2", [{"content-type", "application/x-www-form-urlencoded"}])
+      assert {:ok, %HTTPResponse{status: 200}} = Mint.request(:post, TestServer.url("/post"), "a=1&b=2", [{"content-type", "application/x-www-form-urlencoded"}])
     end
   end
 end