+++ test/assent/strategies/azure_ad_test.exs
@@ -22,7 +22,11 @@ defmodule Assent.Strategy.AzureADTest do
   @user %{
     "name" => "Abe Lincoln",
     "preferred_username" => "AbeLi@microsoft.com",
-    "sub" => "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ"
+    "sub" => "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
+    "aio" => "Df2UVXL1ix!lMCWMSOJBcFatzcGfvFGhjKv8q5g0x732dR5MB5BisvGQO7YWByjd8iQDLq!eGbIDakyp5mnOrcdqHeYSnltepQmRp6AIZ8jY",
+    "oid" => "00000000-0000-0000-66f3-3332eca7ea81",
+    "tid" => "3338040d-6c67-4c5b-b112-36a304b66dad",
+    "ver" => "2.0"
   }
 
   test "authorize_url/2", %{config: config} do
@@ -32,7 +36,7 @@ defmodule Assent.Strategy.AzureADTest do
     assert url =~ "response_mode=form_post"
   end
 
-  test "callback/2", %{config: config, callback_params: params, bypass: bypass} do
+  test "callback/2", %{config: config, callback_params: params} do
     openid_config =
       config[:openid_configuration]
       |> Map.put("issuer", "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0")
@@ -41,8 +45,8 @@ defmodule Assent.Strategy.AzureADTest do
     session_params = Map.put(config[:session_params], :nonce, "123523")
     config         = Keyword.merge(config, openid_configuration: openid_config, tenant_id: "9188040d-6c67-4c5b-b112-36a304b66dad", client_id: "6cb04018-a3f5-46a7-b995-940c78f5aef3", session_params: session_params)
 
-    [key | _rest] = expect_oidc_jwks_uri_request(bypass)
-    expect_oidc_access_token_request(bypass, id_token_opts: [claims: @id_token_claims, kid: key["kid"]])
+    [key | _rest] = expect_oidc_jwks_uri_request()
+    expect_oidc_access_token_request(id_token_opts: [claims: @id_token_claims, kid: key["kid"]])
 
     assert {:ok, %{user: user}} = AzureAD.callback(config, params)
     assert user == @user