+++ test/assent/strategies/oidc_test.exs
@@ -4,21 +4,21 @@ defmodule Assent.Strategy.OIDCTest do
   alias Assent.{JWTAdapter.AssentJWT, RequestError, Strategy.OIDC}
 
   describe "authorize_url/2" do
-    test "generates url and state", %{config: config, bypass: bypass} do
+    test "generates url and state", %{config: config} do
       assert {:ok, %{url: url, session_params: %{state: state}}} = OIDC.authorize_url(config)
 
       refute is_nil(state)
-      assert url =~ "http://localhost:#{bypass.port}/oauth/authorize?client_id=id&redirect_uri=http%3A%2F%2Flocalhost%3A4000%2Fauth%2Fcallback&response_type=code&scope=openid&state=#{state}"
+      assert url =~ TestServer.url("/oauth/authorize?client_id=id&redirect_uri=http%3A%2F%2Flocalhost%3A4000%2Fauth%2Fcallback&response_type=code&scope=openid&state=#{state}")
     end
 
-    test "can add nonce", %{config: config, bypass: bypass} do
+    test "can add nonce", %{config: config} do
       assert {:ok, %{url: url, session_params: %{state: state, nonce: nonce}}} =
         config
         |> Keyword.put(:nonce, "n-0S6_WzA2Mj")
         |> OIDC.authorize_url()
 
       assert nonce == "n-0S6_WzA2Mj"
-      assert url =~ "http://localhost:#{bypass.port}/oauth/authorize?client_id=id&nonce=n-0S6_WzA2Mj&redirect_uri=http%3A%2F%2Flocalhost%3A4000%2Fauth%2Fcallback&response_type=code&scope=openid&state=#{state}"
+      assert url =~ TestServer.url("/oauth/authorize?client_id=id&nonce=n-0S6_WzA2Mj&redirect_uri=http%3A%2F%2Flocalhost%3A4000%2Fauth%2Fcallback&response_type=code&scope=openid&state=#{state}")
     end
   end
 
@@ -34,25 +34,43 @@ defmodule Assent.Strategy.OIDCTest do
       assert OIDC.callback(config, params) == {:error, "`token_endpoint` not found in OpenID configuration"}
     end
 
-    test "with invalid id_token", %{config: config, callback_params: params, bypass: bypass} do
-      expect_oidc_access_token_request(bypass, [id_token: "invalid"])
+    test "with invalid id_token", %{config: config, callback_params: params} do
+      expect_oidc_access_token_request([id_token: "invalid"])
 
       assert OIDC.callback(config, params) == {:error, "The ID Token is not a valid JWT"}
     end
 
-    @user_claims %{sub: "1", name: "Dan Schultzer", email: "foo@example.com", email_verified: true}
-    @user %{"email" => "foo@example.com", "name" => "Dan Schultzer", "sub" => "1", "email_verified" => true}
+    @user_claims %{sub: "1", name: "Dan Schultzer", email: "foo@example.com", email_verified: true, "http://localhost:4000/additional": "info"}
+    @user %{"email" => "foo@example.com", "name" => "Dan Schultzer", "sub" => "1", "email_verified" => true, "http://localhost:4000/additional" => "info"}
 
-    test "with `client_secret_basic` authentication method", %{config: config, callback_params: params, bypass: bypass} do
-      expect_oidc_access_token_request(bypass, [id_token_opts: [claims: @user_claims, iss: "http://localhost"]], fn conn, _params ->
-        assert [{"authorization", "Basic " <> token} | _rest] = conn.req_headers
+    test "with invalid authentication method", %{config: config, callback_params: params} do
+      config = Keyword.put(config, :client_authentication_method, "invalid")
+
+      assert OIDC.callback(config, params) == {:error, "Invalid client authentication method: invalid"}
+    end
+
+    test "with unsupported authentication method", %{config: config, callback_params: params} do
+      openid_configuration =
+        Map.put(config[:openid_configuration], "token_endpoint_auth_methods_supported", ["private_key_jwt"])
+
+      config =
+        config
+        |> Keyword.put(:client_authentication_method, "client_secret_basic")
+        |> Keyword.put(:openid_configuration, openid_configuration)
+
+      assert OIDC.callback(config, params) == {:error, "Unsupported client authentication method: client_secret_basic"}
+    end
+
+    test "with `client_secret_basic` authentication method", %{config: config, callback_params: params} do
+      expect_oidc_access_token_request([id_token_opts: [claims: @user_claims, iss: "http://localhost"]], fn conn, _params ->
+        assert ["Basic " <> token] = Plug.Conn.get_req_header(conn, "authorization")
         assert [client_id, client_secret] = String.split(Base.url_decode64!(token, padding: false), ":")
 
         assert client_id == config[:client_id]
         assert client_secret == config[:client_secret]
       end)
 
-      expect_oidc_jwks_uri_request(bypass, count: 1)
+      expect_oidc_jwks_uri_request(count: 1)
 
       assert {:ok, %{user: user, token: token}} = OIDC.callback(config, params)
       assert user == @user
@@ -100,11 +118,9 @@ defmodule Assent.Strategy.OIDCTest do
       -----END PUBLIC KEY-----
       """
 
-    test "with `private_key_jwt` authentication method", %{config: config, callback_params: params, bypass: bypass} do
+    test "with `private_key_jwt` authentication method", %{config: config, callback_params: params} do
       openid_configuration =
-        config[:openid_configuration]
-        |> Map.put("client_authentication_method", "private_key_jwt")
-        |> Map.put("token_endpoint_auth_methods_supported", ["private_key_jwt"])
+        Map.put(config[:openid_configuration], "token_endpoint_auth_methods_supported", ["private_key_jwt"])
 
       config =
         config
@@ -113,18 +129,20 @@ defmodule Assent.Strategy.OIDCTest do
         |> Keyword.put(:private_key, @private_rsa_key)
         |> Keyword.put(:private_key_id, "key_id")
 
-      expect_oidc_access_token_request(bypass, [id_token_opts: [claims: @user_claims, iss: "http://localhost"]], fn _conn, params ->
+      url = TestServer.url()
+
+      expect_oidc_access_token_request([id_token_opts: [claims: @user_claims, iss: "http://localhost"]], fn _conn, params ->
         assert {:ok, jwt} = AssentJWT.verify(params["client_assertion"], @public_rsa_key, json_library: Jason)
         assert jwt.header["alg"] == "RS256"
         assert jwt.header["typ"] == "JWT"
         assert jwt.header["kid"] == "key_id"
         assert jwt.claims["iss"] == "id"
         assert jwt.claims["sub"] == "id"
-        assert jwt.claims["aud"] == "http://localhost:#{bypass.port}"
+        assert jwt.claims["aud"] == url
         assert jwt.claims["exp"] > DateTime.to_unix(DateTime.utc_now())
       end)
 
-      expect_oidc_jwks_uri_request(bypass, count: 1)
+      expect_oidc_jwks_uri_request(count: 1)
 
       assert {:ok, %{user: user, token: token}} = OIDC.callback(config, params)
       assert user == @user
@@ -133,40 +151,67 @@ defmodule Assent.Strategy.OIDCTest do
   end
 
   describe "callback/2 with dynamic OpenID configuration" do
-    setup %{config: config, bypass: bypass} do
+    setup %{config: config} do
       config = Keyword.delete(config, :openid_configuration)
 
       openid_config = %{
         "token_endpoint" => "/dynamic/token/path",
-        "jwks_uri" => "http://localhost:#{bypass.port}/jwks_uri.json",
+        "jwks_uri" => TestServer.url("/jwks_uri.json"),
         "issuer" => config[:site]
       }
 
       {:ok, config: config, openid_config: openid_config}
     end
 
-    test "with invalid id_token", %{config: config, openid_config: openid_config, callback_params: params, bypass: bypass} do
-      expect_openid_config_request(bypass, openid_config)
+    test "with unreachable openid config url", %{config: config, callback_params: params} do
+      openid_config_url = TestServer.url("/.well-known/openid-configuration")
+      TestServer.stop()
 
-      expect_oidc_access_token_request(bypass, uri: "/dynamic/token/path", id_token: "invalid")
+      assert {:error, %RequestError{} = error} = OIDC.callback(config, params)
+      assert error.error == :unreachable
+      assert error.message =~ "Server was unreachable with Assent.HTTPAdapter.Httpc."
+      assert error.message =~ "{:failed_connect"
+      assert error.message =~ "URL: #{openid_config_url}"
+    end
+
+    test "with unexpected openid config url response", %{config: config, openid_config: openid_config, callback_params: params} do
+      expect_openid_config_request(openid_config, status_code: 201)
+
+      assert {:error, %RequestError{} = error} = OIDC.callback(config, params)
+      assert error.error == :unexpected_response
+      assert error.message =~ "An unexpected success response was received:"
+    end
+
+    test "with 404 openid config url", %{config: config, openid_config: openid_config, callback_params: params} do
+      expect_openid_config_request(openid_config, status_code: 404)
+
+      assert {:error, %RequestError{} = error} = OIDC.callback(config, params)
+      assert error.error == :invalid_server_response
+      assert error.message =~ "Server responded with status: 404"
+    end
+
+    test "with invalid id_token", %{config: config, openid_config: openid_config, callback_params: params} do
+      expect_openid_config_request(openid_config)
+
+      expect_oidc_access_token_request(uri: "/dynamic/token/path", id_token: "invalid")
 
       assert OIDC.callback(config, params) == {:error, "The ID Token is not a valid JWT"}
     end
 
-    test "with missing id_token", %{config: config, openid_config: openid_config, callback_params: params, bypass: bypass} do
-      expect_openid_config_request(bypass, openid_config)
+    test "with missing id_token", %{config: config, openid_config: openid_config, callback_params: params} do
+      expect_openid_config_request(openid_config)
 
-      expect_oidc_access_token_request(bypass, uri: "/dynamic/token/path", params: %{access_token: "access_token", renewal_token: "renewal_token"})
+      expect_oidc_access_token_request(uri: "/dynamic/token/path", params: %{access_token: "access_token", renewal_token: "renewal_token"})
 
       assert OIDC.callback(config, params) ==  {:error, "The `id_token` key not found in token params, only found these keys: access_token, renewal_token"}
     end
 
-    test "with valid id_token", %{config: config, openid_config: openid_config, callback_params: params, bypass: bypass} do
-      expect_openid_config_request(bypass, openid_config)
+    test "with valid id_token", %{config: config, openid_config: openid_config, callback_params: params} do
+      expect_openid_config_request(openid_config)
 
-      expect_oidc_access_token_request(bypass, uri: "/dynamic/token/path")
+      expect_oidc_access_token_request(uri: "/dynamic/token/path")
 
-      expect_oidc_jwks_uri_request(bypass, count: 1)
+      expect_oidc_jwks_uri_request(count: 1)
 
       assert {:ok, %{user: user}} = OIDC.callback(config, params)
       assert user == %{"sub" => "1"}
@@ -195,12 +240,12 @@ defmodule Assent.Strategy.OIDCTest do
       {:ok, id_token: id_token, config: config}
     end
 
-    test "with no openid configuration", %{config: config, id_token: id_token, bypass: bypass} do
+    test "with no openid configuration", %{config: config, id_token: id_token} do
       config = Keyword.delete(config, :openid_configuration)
 
-      expect_openid_config_request(bypass, [], status_code: 500)
+      expect_openid_config_request([], status_code: 500)
 
-      assert {:error, %Assent.RequestError{error: :invalid_server_response}} = OIDC.validate_id_token(config, id_token)
+      assert {:error, %RequestError{error: :invalid_server_response}} = OIDC.validate_id_token(config, id_token)
     end
 
     test "with no `:client_id`", %{config: config, id_token: id_token} do
@@ -254,7 +299,58 @@ defmodule Assent.Strategy.OIDCTest do
     test "with invalid `aud` in id_token", %{config: config} do
       id_token = gen_id_token(alg: "HS256", claims: %{"aud" => "invalid"})
 
-      assert OIDC.validate_id_token(config, id_token) == {:error, "Invalid audience \"invalid\" in ID Token"}
+      assert OIDC.validate_id_token(config, id_token) == {:error, "`:client_id` not in audience [\"invalid\"] in ID Token"}
+    end
+
+    test "with case insensitive `aud` in id_token", %{config: config} do
+      id_token = gen_id_token(alg: "HS256", claims: %{"aud" => "ID"})
+
+      assert OIDC.validate_id_token(config, id_token) == {:error, "`:client_id` not in audience [\"ID\"] in ID Token"}
+    end
+
+    test "with `aud` list with only client id in id_token", %{config: config} do
+      id_token = gen_id_token(alg: "HS256", claims: %{"aud" => ~w(id)})
+
+      assert {:ok, _} = OIDC.validate_id_token(config, id_token)
+    end
+
+    test "with `aud` list in id_token no trusted_audiences config", %{config: config} do
+      id_token = gen_id_token(alg: "HS256", claims: %{"aud" => ~w(id aud1 aud2)})
+
+      assert OIDC.validate_id_token(config, id_token) == {:error, "Untrusted audience(s) [\"aud1\", \"aud2\"] in ID Token"}
+    end
+
+    test "with `aud` list with missing client id in id_token", %{config: config} do
+      id_token = gen_id_token(alg: "HS256", claims: %{"aud" => ~w(aud1 aud2)})
+      config = Keyword.put(config, :trusted_audiences, ~w(aud1 aud2))
+
+      assert OIDC.validate_id_token(config, id_token) == {:error, "`:client_id` not in audience [\"aud1\", \"aud2\"] in ID Token"}
+    end
+
+    test "with `aud` list with untrusted audiences in id_token", %{config: config} do
+      id_token = gen_id_token(alg: "HS256", claims: %{"aud" => ~w(id aud1 aud2)})
+      config = Keyword.put(config, :trusted_audiences, ~w(aud1))
+
+      assert OIDC.validate_id_token(config, id_token) == {:error, "Untrusted audience(s) [\"aud2\"] in ID Token"}
+    end
+
+    test "with `aud` list in id_token", %{config: config} do
+      id_token = gen_id_token(alg: "HS256", claims: %{"aud" => ~w(id aud1 aud2)})
+      config = Keyword.put(config, :trusted_audiences, ~w(aud1 aud2))
+
+      assert {:ok, _} = OIDC.validate_id_token(config, id_token)
+    end
+
+    test "with invalid `azp`", %{config: config} do
+      id_token = gen_id_token(alg: "HS256", claims: %{"azp" => "invalid"})
+
+      assert OIDC.validate_id_token(config, id_token) == {:error, "Invalid authorized party \"invalid\" in ID Token"}
+    end
+
+    test "with valid `azp`", %{config: config} do
+      id_token = gen_id_token(alg: "HS256", claims: %{"azp" => "id"})
+
+      assert {:ok, _} = OIDC.validate_id_token(config, id_token)
     end
     
     test "with valid `aud` list in id_token", %{config: config} do
@@ -340,18 +436,37 @@ defmodule Assent.Strategy.OIDCTest do
       assert OIDC.validate_id_token(config, id_token) == {:error, "`jwks_uri` not found in OpenID configuration"}
     end
 
-    test "with 404 `jwks_uri` url", %{config: config, id_token: id_token, bypass: bypass} do
-      Bypass.expect_once(bypass, "GET", "/jwks_uri.json", fn conn ->
+    test "with unreachable `jwk_uri` url",  %{config: config, id_token: id_token} do
+      jwks_uri_url = TestServer.url("/jwks_uri.json")
+      TestServer.stop()
+
+      assert {:error, %RequestError{} = error} = OIDC.validate_id_token(config, id_token)
+      assert error.error == :unreachable
+      assert error.message =~ "Server was unreachable with Assent.HTTPAdapter.Httpc."
+      assert error.message =~ "{:failed_connect"
+      assert error.message =~ "URL: #{jwks_uri_url}"
+    end
+
+    test "with unexpected `jwk_uri` url response", %{config: config, id_token: id_token} do
+      expect_oidc_jwks_uri_request(status_code: 201)
+
+      assert {:error, %RequestError{} = error} = OIDC.validate_id_token(config, id_token)
+      assert error.error == :unexpected_response
+      assert error.message =~ "An unexpected success response was received:"
+    end
+
+    test "with 404 `jwks_uri` url", %{config: config, id_token: id_token} do
+      TestServer.add("/jwks_uri.json", via: :get, to: fn conn ->
         Plug.Conn.send_resp(conn, 404, "")
       end)
 
-      assert {:error, %Assent.RequestError{} = error} = OIDC.validate_id_token(config, id_token)
+      assert {:error, %RequestError{} = error} = OIDC.validate_id_token(config, id_token)
       assert error.error == :invalid_server_response
       assert error.message =~ "Server responded with status: 404"
     end
 
-    test "with missing keys in `jwks_uri` url", %{config: config, id_token: id_token, bypass: bypass} do
-      Bypass.expect_once(bypass, "GET", "/jwks_uri.json", fn conn ->
+    test "with missing keys in `jwks_uri` url", %{config: config, id_token: id_token} do
+      TestServer.add("/jwks_uri.json", via: :get, to: fn conn ->
         conn
         |> Plug.Conn.put_resp_content_type("application/json")
         |> Plug.Conn.send_resp(200, "{}")
@@ -360,32 +475,32 @@ defmodule Assent.Strategy.OIDCTest do
       assert OIDC.validate_id_token(config, id_token) == {:error, "No keys found in `jwks_uri` provider configuration"}
     end
 
-    test "with no `kid` in header and multiple keys fetched from `jwks_uri` url", %{config: config, id_token: id_token, bypass: bypass} do
-      expect_oidc_jwks_uri_request(bypass)
+    test "with no `kid` in header and multiple keys fetched from `jwks_uri` url", %{config: config, id_token: id_token} do
+      expect_oidc_jwks_uri_request()
 
       assert OIDC.validate_id_token(config, id_token) == {:error, "Multiple public keys found in provider configuration and no `kid` value in ID Token"}
     end
 
-    test "with no `kid` in header and single key fetched from `jwks_uri` url", %{config: config, id_token: id_token, bypass: bypass} do
-      expect_oidc_jwks_uri_request(bypass, count: 1)
+    test "with no `kid` in header and single key fetched from `jwks_uri` url", %{config: config, id_token: id_token} do
+      expect_oidc_jwks_uri_request(count: 1)
 
       assert {:ok, jwt} = OIDC.validate_id_token(config, id_token)
       assert jwt.verified?
       assert jwt.claims["sub"] == "1"
     end
 
-    test "with no matching `kid` in keys fetched from `jwks_uri` url", %{config: config, bypass: bypass} do
+    test "with no matching `kid` in keys fetched from `jwks_uri` url", %{config: config} do
       id_token = gen_id_token(kid: "invalid")
 
-      expect_oidc_jwks_uri_request(bypass)
+      expect_oidc_jwks_uri_request()
 
       assert OIDC.validate_id_token(config, id_token) == {:error, "No keys found for the `kid` value \"invalid\" provided in ID Token"}
     end
 
-    test "with matching `kid` in keys fetched from `jwks_uri` url", %{config: config, bypass: bypass} do
+    test "with matching `kid` in keys fetched from `jwks_uri` url", %{config: config} do
       id_token = gen_id_token(kid: "key-1")
 
-      expect_oidc_jwks_uri_request(bypass)
+      expect_oidc_jwks_uri_request()
 
       assert {:ok, jwt} = OIDC.validate_id_token(config, id_token)
       assert jwt.verified?
@@ -402,11 +517,11 @@ defmodule Assent.Strategy.OIDCTest do
       {:ok, config: config, access_token: access_token}
     end
 
-    test "with no openid configuration", %{config: config, access_token: access_token, bypass: bypass} do
+    test "with no openid configuration", %{config: config, access_token: access_token} do
       config = Keyword.delete(config, :openid_configuration)
-      expect_openid_config_request(bypass, [], status_code: 500)
+      expect_openid_config_request([], status_code: 500)
 
-      assert {:error, %Assent.RequestError{error: :invalid_server_response}} = OIDC.fetch_userinfo(config, access_token)
+      assert {:error, %RequestError{error: :invalid_server_response}} = OIDC.fetch_userinfo(config, access_token)
     end
 
     test "with missing `userinfo_endpoint` in OpenID configuration", %{config: config, access_token: access_token} do
@@ -427,35 +542,43 @@ defmodule Assent.Strategy.OIDCTest do
       assert error.message =~ "URL: http://localhost:8888/userinfo"
     end
 
-    test "with unauthorized `userinfo_endpoint`", %{config: config, access_token: access_token, bypass: bypass} do
-      expect_oidc_userinfo_request(bypass, %{"error" => "Unauthorized"}, status_code: 401)
+    test "with unexpected `userinfo_endpoint` url response", %{config: config, access_token: access_token} do
+      expect_oidc_userinfo_request(gen_id_token(alg: "HS256"), status_code: 201)
+
+      assert {:error, %RequestError{} = error} = OIDC.fetch_userinfo(config, access_token)
+      assert error.error == :unexpected_response
+      assert error.message =~ "An unexpected success response was received:"
+    end
+
+    test "with unauthorized `userinfo_endpoint`", %{config: config, access_token: access_token} do
+      expect_oidc_userinfo_request(%{"error" => "Unauthorized"}, status_code: 401)
 
       assert {:error, %RequestError{} = error} = OIDC.fetch_userinfo(config, access_token)
       assert error.message == "Unauthorized token"
       refute error.error
     end
 
-    test "with jwt response with invalid signature", %{config: config, access_token: access_token, bypass: bypass} do
+    test "with jwt response with invalid signature", %{config: config, access_token: access_token} do
       [header, payload, _signature] = String.split(gen_id_token(alg: "HS256"), ".")
-      expect_oidc_userinfo_request(bypass, "#{header}.#{payload}.invalid")
+      expect_oidc_userinfo_request("#{header}.#{payload}.invalid")
 
       assert OIDC.fetch_userinfo(config, access_token) == {:error, "Invalid JWT signature for ID Token"}
     end
 
-    test "with jwt response", %{config: config, access_token: access_token, bypass: bypass} do
-      expect_oidc_userinfo_request(bypass, gen_id_token(alg: "HS256"))
+    test "with jwt response", %{config: config, access_token: access_token} do
+      expect_oidc_userinfo_request(gen_id_token(alg: "HS256"))
 
       assert {:ok, %{"sub" => "1"}} = OIDC.fetch_userinfo(config, access_token)
     end
 
-    test "with missing `sub` in userinfo claims", %{config: config, access_token: access_token, bypass: bypass} do
-      expect_oidc_userinfo_request(bypass, Map.delete(@user_claims, :sub))
+    test "with missing `sub` in userinfo claims", %{config: config, access_token: access_token} do
+      expect_oidc_userinfo_request(Map.delete(@user_claims, :sub))
 
       assert OIDC.fetch_userinfo(config, access_token) == {:error, "`sub` not in userinfo response"}
     end
 
-    test "with different `sub` in userinfo claims", %{config: config, access_token: access_token, bypass: bypass} do
-      expect_oidc_userinfo_request(bypass, Map.put(@user_claims, :sub, "2"))
+    test "with different `sub` in userinfo claims", %{config: config, access_token: access_token} do
+      expect_oidc_userinfo_request(Map.put(@user_claims, :sub, "2"))
 
       assert OIDC.fetch_userinfo(config, access_token) == {:error, "`sub` in userinfo response not the same as in ID Token"}
     end