+++ CHANGELOG.md
@@ -1,5 +1,58 @@
 # Changelog
 
+## v0.2.3 (2023-03-23)
+
+* Removed `:castore` version requirement
+* `Assent.Strategy.Httpc.request/5` raises error when SSL certificate can't be validated
+
+## v0.2.2 (2023-02-27)
+
+* Fixed bug to handle 201 success response
+* `Assent.Strategy.OIDC` now has support for multiple audiences
+* `Assent.Strategy.OIDC` now permits any auth method if no `token_endpoint_auth_methods_supported` specified
+* `Assent.Strategy.Linkedin` added
+
+## v0.2.1 (2022-09-15)
+
+* Default to using `Jason` instead of `Poison` for JSON parsing
+* Fixed `Bitwise` warning when running on Elixir 1.14
+
+## v0.2.0 (2022-03-01)
+
+**This release consists of breaking changes.**
+
+In previous `Assent.Strategy.Slack` strategy, the `sub` user id field consisted of `{SUB}-{TEAM}`. Slack has migrated to OpenID Connect, and the response has been conformed to OIDC. The `sub` will now only consists of the `sub` id, and not include team id. To succesfullly migrate to this release all slack identity records storing the `sub` user id field has to be updated.
+
+If you wish to continue using the previous `sub` user id a custom OIDC strategy can be used instead:
+
+```elixir
+defmodule Slack do
+  use Assent.Strategy.OIDC.Base
+
+  alias Assent.Strategy.Slack
+
+  defdelegate default_config(config), to: Slack
+
+  def normalize(config, user) do
+    user = Map.put(user, "sub", "#{user["https://slack.com/user_id"]}-#{user["https://slack.com/team_id"]}")
+
+    Slack.normalize(config, user)
+  end
+end
+```
+
+* `Assent.Strategy.OIDC.fetch_user/2` now removes the ID token specific keys from the user claims instead of normalizing
+* `Assent.Strategy.OIDC.Base` now adds `normalize/2` to the macro that will include the full user claims in the user params
+* `Assent.Strategy.Slack` now uses OpenID connect instead of legacy OAuth 2.0, please note that the `sub` value may have changed
+
+## v0.1.28 (2021-09-30)
+
+* `Assent.Strategy.OIDC` bug fixed so it handles unreachable urls correctly
+
+## v0.1.27 (2021-08-21)
+
+* `Assent.Strategy.OIDC` bug fixed for `normalize/2` macro callback
+
 ## v0.1.26 (2021-05-27)
 
 * `Assent.constant_time_compare/2` no longer outputs a deprecation warning for OTP 24